윤주 231213 product 작업

+++ src/main/java/com/book/book_project/dto/ReviewInterface.java
@@ -6,7 +6,7 @@ import com.book.book_project.entity.ProductEntity;
 import java.sql.Timestamp;
 
 public interface ReviewInterface {
-	Long getReviewseq();
+	int getReviewseq();
 
 	ProductEntity getBookid();
