+++ src/main/java/com/book/book_project/entity/repository/ReviewRepository.java
@@ -10,10 +10,10 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface ReviewRepository extends JpaRepository<ReviewEntity, Long> {
+public interface ReviewRepository extends JpaRepository<ReviewEntity, Integer> {
 @Query(value="select * from tbl_review where bookid=:bookid order by reviewseq desc", nativeQuery = true)
-	List<ReviewInterface> reviewView(@Param("bookid") Long bookid);
+	List<ReviewInterface> reviewView(@Param("bookid") int bookid);
 
 @Query(value="select * from tbl_review where bookid=:bookid order by reviewseq desc", nativeQuery = true)
-	List<ReviewDTO> reviewView1(@Param("bookid") Long bookid);
+	List<ReviewDTO> reviewView1(@Param("bookid") int bookid);
         }
